Removed extra buffer in SBOBJlookup and USloadObject. Now loads directly into AOA. Added SBOBJload and changed SBOBJlookup, USloadObject for this.
Tried to get sun4s to compile after EC's changes to the type of thePC in
various places... He'll get it back now...

Betarun version number: 46


git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@12875 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/C/betaerror.c
@@ -95,7 +95,7 @@ void BetaError(BetaErr err, Object *theObj)
 #ifdef NEWRUN
       DEBUG_CODE(fprintf(output, "          SP=0x%x,\n", SP));
       DEBUG_CODE(fprintf(output, "          thePC=0x%x", (int)thePC));
-      DEBUG_CODE(PrintCodeAddress((long)thePC));
+      DEBUG_CODE(PrintCodeAddress(thePC));
       DEBUG_CODE(fprintf(output, ")\n"));
       switch(err){
       case StopCalledErr /* called via FailureExit in betaenv.o */:
@@ -143,7 +143,7 @@ void BetaError(BetaErr err, Object *theObj)
       thePC = (long *) ((RegWin *)FramePointer)->i7;
       DEBUG_CODE(fprintf(output, "          StackEnd=0x%x,\n", (int)StackEnd));
       DEBUG_CODE(fprintf(output, "          thePC=0x%x", (int)thePC));
-      DEBUG_CODE(PrintCodeAddress((long)thePC));
+      DEBUG_CODE(PrintCodeAddress(thePC));
       DEBUG_CODE(fprintf(output, ")\n"));
 #endif /* sparc */
 
@@ -161,7 +161,7 @@ void BetaError(BetaErr err, Object *theObj)
 #error Find out Stack End for hppa without Reference Stack
 #endif /* UseRefStack */
       DEBUG_CODE(fprintf(output, "thePC=0x%x", (int)thePC));
-      DEBUG_CODE(PrintCodeAddress((long)thePC));
+      DEBUG_CODE(PrintCodeAddress(thePC));
       DEBUG_CODE(fprintf(output, ")\n"));
 #endif /* hppa */
